GB Video: Increase palette entry width to 24 bits

+++ include/mgba/internal/gb/video.h
@@ -144,7 +144,7 @@ void GBVideoWriteLYC(struct GBVideo* video, uint8_t value);
 void GBVideoWritePalette(struct GBVideo* video, uint16_t address, uint8_t value);
 void GBVideoSwitchBank(struct GBVideo* video, uint8_t value);
 
-void GBVideoSetPalette(struct GBVideo* video, unsigned index, uint16_t color);
+void GBVideoSetPalette(struct GBVideo* video, unsigned index, uint32_t color);
 
 struct GBSerializedState;
 void GBVideoSerialize(const struct GBVideo* video, struct GBSerializedState* state);
